+++ pydidas/apps/parsers/__init__.py
@@ -19,7 +19,7 @@ command line with command-line calling arguments.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2021-2022, Malte Storm, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023, Malte Storm, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0"
 __maintainer__ = "Malte Storm"
 __status__ = "Development"